% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{annotateCells}
\alias{annotateCells}
\title{Annotate cells in a Seurat object using SCINA}
\usage{
annotateCells(
  seur,
  org = c("HUMAN", "MOUSE", "RAT"),
  tissue = "ALL",
  id.type = c("Symbol", "Ensembl"),
  output_path = paste0(output_dir, "/UMAP_SCINA.png")
)
}
\arguments{
\item{seur}{A Seurat object containing the single-cell RNA-seq data.}

\item{org}{Character string specifying the organism. Must be one of "HUMAN", "MOUSE", or "RAT".}

\item{tissue}{Character string specifying the tissue type. Default is "ALL".}

\item{id.type}{Character string specifying the gene ID type. Must be either "Symbol" or "Ensembl".}

\item{output_path}{Character string specifying the path to save the UMAP plot.
Default is "UMAP_SCINA.png" in the output_dir.}
}
\value{
A Seurat object with added SCINA annotations in the meta.data slot.
}
\description{
This function annotates cells in a Seurat object using the SCINA algorithm.
It processes the data, selects appropriate markers based on the organism and tissue,
and applies the SCINA algorithm for cell type annotation.
}
\examples{
seur <- annotateCells(seur, org = "HUMAN", tissue = "Liver", id.type = "Symbol")

}
