% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellCycleGenes_HUMAN}
\alias{CellCycleGenes_HUMAN}
\title{Cell cycle genes data for human}
\format{
A data frame with columns for cell cycle phase, gene ID, and symbol
}
\source{
Your source information here
}
\usage{
CellCycleGenes_HUMAN
}
\description{
Cell cycle genes data for human
}
\keyword{datasets}
