% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellCycleGenes_MOUSE}
\alias{CellCycleGenes_MOUSE}
\title{Cell cycle genes data for mouse}
\format{
A data frame with columns for cell cycle phase, gene ID, and symbol
}
\source{
##TO DO: detail ensembl process here, same as commentedin function
}
\usage{
CellCycleGenes_MOUSE
}
\description{
Cell cycle genes data for mouse
}
\keyword{datasets}
