% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CellCycleGenes_RAT}
\alias{CellCycleGenes_RAT}
\title{Cell cycle genes data for rat}
\format{
A data frame with columns for cell cycle phase, gene ID, and symbol
}
\source{
Your source information here
}
\usage{
CellCycleGenes_RAT
}
\description{
Cell cycle genes data for rat
}
\keyword{datasets}
