% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm_human}
\alias{cm_human}
\title{Cell markers data for human}
\format{
A list with cell type markers for various human tissues
}
\source{
Your source information here
}
\usage{
cm_human
}
\description{
Cell markers data for human
}
\keyword{datasets}
