% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ExploreCluster}
\alias{ExploreCluster}
\title{ExploreCluster}
\usage{
ExploreCluster(
  seur,
  cluster,
  runPCAUMAP = FALSE,
  ucomps = 50,
  comps = 100,
  verb = T,
  k = 20
)
}
\arguments{
\item{seur}{A Seurat object of single-cell data with the 'integrated' assay.}

\item{cluster}{An integer or name that can be found by the subset function.}

\item{runPCAUMAP}{(Optional) A boolean indicating whether to rerun PCA and UMAP. Default is FALSE. Not recommended to set it to TRUE unless results with FALSE not satisfactory.}

\item{ucomps}{(Optional) Number of PCA components to use in UMAP. Default is 50.}

\item{comps}{(Optional) Number of components for PCA if ran. Default is 100.}

\item{verb}{(Optional) A boolean to decide whether to print messages during PCA. Default is TRUE.}

\item{k}{(Optional) Number of nearest neighbors for UMAP. Default is 20.}
}
\value{
A modified Seurat object of the selected cluster divided into subclusters (e.g., 2.1, 2.2, 2.3).
}
\description{
This function takes a Seurat object and a cluster number or name, and
optionally reruns PCA and UMAP on the selected cluster. Note that despite the
option being offered, it is NOT recommended and it seems that using the
clustering from the original data leads to better results.
It returns a subsetted Seurat object with numbered subclusters
branching from the original name separated by a dot.
Follow up with RemapClusters to put them back in original object
recommended, before or after identifying and renaming them.
}
\examples{
seur <- ExploreCluster(seur, 2)
seur <- ExploreCluster(seur, 2), with no additional subclustering due to
default runPCAUMAP to false
returns subset seurat object with cells in cluster 2 divided in 2.1, 2.2, 2.3
}
