% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filterCells}
\alias{filterCells}
\title{Filter cells based on quality metrics}
\usage{
filterCells(
  seur,
  mad.coeff = 3,
  pass = -1,
  org = "HUMAN",
  output_path = output_dir
)
}
\arguments{
\item{seur}{A Seurat object}

\item{mad.coeff}{Coefficient for median absolute deviation, default is 3}

\item{pass}{An identifier for the number of times this function is running, used for plotting}

\item{org}{The organism the data is from ("HUMAN", "MOUSE", or "RAT")}

\item{output_path}{Directory to save output plots, default is output_dir}
}
\value{
A filtered Seurat object
}
\description{
This function filters cells based on percentage of mitochondrial transcripts,
number of genes, and number of transcripts detected.
A median absolute deviation approach, letting the distribution of the data define cutoffs, is chosen
}
\examples{
filtered_seurat <- filterCells(seurat_obj, mad.coeff = 3, pass = 1, org = "HUMAN")
}
