% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{findOptimalResolution}
\alias{findOptimalResolution}
\title{Find Optimal Resolution for Clustering}
\usage{
findOptimalResolution(seur, pcs, output_path = output_dir)
}
\arguments{
\item{seur}{A Seurat object.}

\item{pcs}{An integer vector specifying which principal components to use
(e.g., 1:50).}

\item{output_path}{A string specifying the directory to save the plot.
Default is \code{output_dir}.}
}
\value{
A modified Seurat object with the winning clustering in
\code{seurat_clusters} and \code{Idents(seur)}.
}
\description{
This function takes a Seurat object and a range of PC components,
calculates the Calinski-Harabasz index at a range of resolutions (Leiden),
and applies the resolution that yields the highest CH index back to the
Seurat object's \code{seurat_clusters} metadata.
A plot of the CH index across resolutions is saved.
}
\examples{
seur <- findOptimalResolution(seur, pcs = 1:50)
}
