% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_necessary_data}
\alias{load_necessary_data}
\title{Load necessary data for SmartQC}
\usage{
load_necessary_data(org)
}
\arguments{
\item{org}{Organism (MOUSE, HUMAN, or RAT) given by user at start}
}
\value{
A list containing necessary marker data for SCINA and necessary cell cycle genes data for vst regress out
}
\description{
Load necessary data for SmartQC
}
