% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MergeClusters}
\alias{MergeClusters}
\title{Merge Clusters}
\usage{
MergeClusters(seur, clusters_to_merge, use_regex = FALSE)
}
\arguments{
\item{seur}{A Seurat object containing the original data.}

\item{clusters_to_merge}{A list of clusters to merge.
If use_regex is TRUE, clusters_to_merge should be a list of vectors, where each vector contains clusters to merge based on regex patterns.
If use_regex is FALSE, clusters_to_merge should be a list of pairs, where each pair contains two clusters to merge.}

\item{use_regex}{A boolean indicating whether to use regex patterns for merging clusters (default is FALSE).
If use_regex is TRUE, clusters will be merged based on regex patterns.
If use_regex is FALSE, specific clusters will be merged as specified in the input list.
Note final behavior is different!!! FALSE and all subclusters will not give you equivalent to regex TRUE}
}
\value{
A modified Seurat object with the specified clusters merged in the "RefinedClusters" metadata column.
}
\description{
This function merges clusters in the RefinedClusters metadata column of a Seurat object.
You can choose between two behaviors: merging based on regex patterns or merging specific clusters.
}
\examples{
# Merge clusters 2.1, 2.2, and 2.3 into a single cluster named 2 using the regex behavior
seur <- MergeClusters(seur, list(c("2.1", "2.2", "2.3")), use_regex = TRUE)

# Merge cluster 2.3 into cluster 2.1 using the specific cluster merging behavior
seur <- MergeClusters(seur, list(c("2.1", "2.3")), use_regex = FALSE)
}
