% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartqc.R
\name{process_sample}
\alias{process_sample}
\title{Process a single sample}
\usage{
process_sample(
  input_dir,
  organism,
  cellcyclegenes,
  output_dir = NULL,
  run_umap = TRUE,
  run_densmap = FALSE,
  run_scina = FALSE,
  umap_neighbors = 20
)
}
\arguments{
\item{input_dir}{Directory containing the input data}

\item{organism}{Organism (MOUSE, HUMAN, or RAT)}

\item{cellcyclegenes}{Cell cycle genes data}

\item{output_dir}{Directory to save output files (optional)}

\item{run_umap}{Whether to run UMAP (default: TRUE)}

\item{run_densmap}{Whether to run DensMAP (default: FALSE)}

\item{run_scina}{Whether to run SCINA (default: FALSE)}

\item{umap_neighbors}{Number of neighbors for UMAP (default: 20)}
}
\value{
A Seurat object containing the processed data
}
\description{
This function processes a single sample of single-cell RNA-seq data.
}
