% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ProcessClusters}
\alias{ProcessClusters}
\title{Process Clusters}
\usage{
ProcessClusters(seur, clusters, runPCAUMAP = rep(FALSE, length(clusters)))
}
\arguments{
\item{seur}{A Seurat object containing all the original data}

\item{clusters}{A vector of integers specifying the clusters to process.}

\item{runPCAUMAP}{(Optional) A boolean vector indicating whether to run PCA and UMAP for each cluster.
Default is FALSE for each cluster. The length of runPCAUMAP must match the length of clusters.}
}
\value{
A modified Seurat object with refined clustering in the "RefinedClusters" metadata column.
}
\description{
This function takes an integrated Seurat object, a vector of clusters to process, and an optional
boolean vector indicating whether to run PCA and UMAP for each cluster. It runs ExploreCluster and
RemapClusters functions serially for each specified cluster, giving you more clusters only in the selected clusters list.
}
\details{
Intended workflow is to use this when initial exploration of gene markers
suggests that some clusters include too many different types of cell
and you must break them down for proper identification.
}
\examples{
clusters_to_process <- c(2, 3, 4)
runPCAUMAP <- c(FALSE, TRUE, FALSE) #note false at all types recommended
seur <- ProcessClusters(seur, clusters_to_process, runPCAUMAP)

}
