% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RemapClusters}
\alias{RemapClusters}
\title{Remap Clusters}
\usage{
RemapClusters(seur, sub, mcluster)
}
\arguments{
\item{seur}{A Seurat object containing all the original data}

\item{sub}{A Seurat object containing the subset clusters to be mapped back into the seur object.}

\item{mcluster}{The cluster integer that was used to create the subset object.}
}
\value{
A modified Seurat object with the subset clusters remapped and added to the "RefinedClusters" metadata column.
}
\description{
Taking your original integrated Seurat object, a sub object from
ExploreCluster and a specific cluster that you already ran ExploreCLuster on,
It creates a new metadata column, "RefinedClusters", putting them back in
original object. In sub_backup you have all the original numeric clusters
}
\details{
If the cluster is a number RefinedClusters will have
original seurat_cluster + . + subcluster number
If it's already a cell type name it's just put into place
Both cases matching by UMIs
}
\examples{
seur <- RemapClusters(seur, sub, 3)


}
