% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartqc.R
\name{run_smartqc}
\alias{run_smartqc}
\title{Run SmartQC pipeline}
\usage{
run_smartqc(
  input = NULL,
  output = NULL,
  root = NULL,
  study = NULL,
  organism = "MOUSE"
)
}
\arguments{
\item{input}{Input directory with raw data (for single sample processing)}

\item{output}{Output file name (for single sample processing)}

\item{root}{Root directory containing Studies folder (for multi-sample processing)}

\item{study}{Study name (for multi-sample processing)}

\item{organism}{Organism (MOUSE, HUMAN, or RAT)}
}
\description{
Run SmartQC pipeline
}
