% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm_rat}
\alias{cm_rat}
\title{Cell markers data for rat}
\format{
A list with cell type markers for various rat tissues
}
\source{
Your source information here
}
\usage{
cm_rat
}
\description{
Cell markers data for rat
}
\keyword{datasets}
