% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_necessary_data}
\alias{load_necessary_data}
\title{Load necessary data for SmartQC}
\usage{
load_necessary_data(org)
}
\arguments{
\item{org}{Organism (MOUSE, HUMAN, or RAT)}
}
\value{
A list containing loaded data
}
\description{
Load necessary data for SmartQC
}
